Refactor utility imports and reorganize test structure
