+++ app.py
@@ -6,7 +6,7 @@ from flask_cors import CORS
 
 from config import Config, limiter
 from routes import register_routes
-from utils import extract_error_message
+from utility import extract_error_message
 
 app = Flask(__name__)
 app.config.from_object(Config)
