+++ tests/test_routes/__init__.py
@@ -1,0 +1,1 @@
+from routes import register_routes
