+++ tests/test_routes/test_home.py
@@ -1,0 +1,8 @@
+from app import app
+
+
+def test_index_route():
+    response = app.test_client().get("/")
+
+    assert response.status_code == 200
+    assert response.data.decode("utf-8") != ""
