+++ utility.py
